Move/delete login files
